% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDVI.R
\name{NDVI}
\alias{NDVI}
\title{NDVI}
\usage{
NDVI(NIR, R)
}
\arguments{
\item{NIR}{a dataframe or matrix contains NIR band values.}

\item{R}{a dataframe or matrix contains red band values, NIR and R should have same class and dimension..}
}
\value{
A matrix or dataframe(depends on the class of NIR and R).
}
\description{
This function calculate NDVI value from input near-infrared (NIR) and red bands images.
If change the input from red band into green or red-edge band values, it will return GNDVI or NDRE values.
}
\references{
De Swaef, T., Maes, W. H., Aper, J., Baert, J., Cougnon, M., Reheul, D., ... & Lootens, P. (2021). Applying RGB-and thermal-based vegetation indices from UAVs for high-throughput field phenotyping of drought tolerance in forage grasses. Remote Sensing, 13(1), 147.
}
